test: add tests for post request params
