+++ test/helpers/fixtures.ts
@@ -1,7 +1,10 @@
 import { SimpleMerkleTree } from '@openzeppelin/merkle-tree';
 import { ethers, upgrades } from 'hardhat';
+import type { CoreRequestTypes, CoreResultTypes } from '../../ts-types';
 import { ONE_DAY_IN_SECONDS } from '../utils/constants';
-import { computeResultLeafHash, computeValidatorLeafHash, deriveResultId, generateDataFixtures } from '../utils/crypto';
+import { NON_ZERO_HASH, SEDA_DATA_TYPES_VERSION } from '../utils/constants';
+import { computeResultLeafHash, computeValidatorLeafHash, deriveResultId } from '../utils/crypto';
+import { deriveRequestId } from '../utils/crypto';
 
 interface DeployOptions {
   requests?: number;
@@ -99,4 +102,51 @@ export async function deployWithOptions(options: DeployOptions) {
   };
 
   return { prover, core, feeManager, data };
+}
+
+export function generateDataFixtures(
+  length: number,
+  resultLength?: number,
+): {
+  requests: CoreRequestTypes.RequestInputsStruct[];
+  results: CoreResultTypes.ResultStruct[];
+} {
+  const requests = Array.from({ length }, (_, i) => ({
+    execProgramId: NON_ZERO_HASH,
+    execInputs: NON_ZERO_HASH,
+    execGasLimit: 10_000_000_000_000n,
+    tallyProgramId: NON_ZERO_HASH,
+    tallyInputs: NON_ZERO_HASH,
+    tallyGasLimit: 10_000_000_000_000n,
+    replicationFactor: 1,
+    consensusFilter: '0x01',
+    gasPrice: 10000000000n,
+    memo: ethers.hexlify(ethers.toUtf8Bytes(`request-${i + 1}`)),
+  }));
+
+  const results = requests.map((request) => {
+    const drId = deriveRequestId(request);
+    const result = resultLength
+      ? `0x${Array.from({ length: resultLength }, () =>
+          Math.floor(Math.random() * 256)
+            .toString(16)
+            .padStart(2, '0'),
+        ).join('')}`
+      : ethers.keccak256(ethers.toUtf8Bytes('SUCCESS'));
+
+    return {
+      version: SEDA_DATA_TYPES_VERSION,
+      drId,
+      consensus: true,
+      exitCode: 0,
+      result,
+      blockHeight: 1,
+      blockTimestamp: Math.floor(Date.now() / 1000) + 3600,
+      gasUsed: 1000000n,
+      paybackAddress: NON_ZERO_HASH,
+      sedaPayload: NON_ZERO_HASH,
+    };
+  });
+
+  return { requests, results };
 }
