+++ test/utils/crypto.ts
@@ -79,50 +79,3 @@ export function computeValidatorLeafHash(validator: string, votingPower: number)
     [SECP256K1_DOMAIN_SEPARATOR, validator, votingPower],
   );
 }
-
-export function generateDataFixtures(
-  length: number,
-  resultLength?: number,
-): {
-  requests: CoreRequestTypes.RequestInputsStruct[];
-  results: CoreResultTypes.ResultStruct[];
-} {
-  const requests = Array.from({ length }, (_, i) => ({
-    execProgramId: NON_ZERO_HASH,
-    execInputs: NON_ZERO_HASH,
-    execGasLimit: 1000000n,
-    tallyProgramId: NON_ZERO_HASH,
-    tallyInputs: NON_ZERO_HASH,
-    tallyGasLimit: 1000000n,
-    replicationFactor: 1,
-    consensusFilter: '0x01',
-    gasPrice: 10000000000n,
-    memo: ethers.hexlify(ethers.toUtf8Bytes(`request-${i + 1}`)),
-  }));
-
-  const results = requests.map((request) => {
-    const drId = deriveRequestId(request);
-    const result = resultLength
-      ? `0x${Array.from({ length: resultLength }, () =>
-          Math.floor(Math.random() * 256)
-            .toString(16)
-            .padStart(2, '0'),
-        ).join('')}`
-      : ethers.keccak256(ethers.toUtf8Bytes('SUCCESS'));
-
-    return {
-      version: SEDA_DATA_TYPES_VERSION,
-      drId,
-      consensus: true,
-      exitCode: 0,
-      result,
-      blockHeight: 1,
-      blockTimestamp: Math.floor(Date.now() / 1000) + 3600,
-      gasUsed: 1000000n,
-      paybackAddress: NON_ZERO_HASH,
-      sedaPayload: NON_ZERO_HASH,
-    };
-  });
-
-  return { requests, results };
-}
